Add files via upload
Here is the most final version of the code.
The correct titles are stored in a csv file (all_titles) and used for the cleaning of titles. There is also a hierarchy applied. For example when there is 'Founder and COO 'it gives the output 'Founder'.
New titles will be added to this csv file.